+++ app/models/contact_request.rb
@@ -1,0 +1,2 @@
+class ContactRequest < Whois::ContactRequest
+end
